tests/wpt/web-platform-tests/html/semantics/popups/popup-focus.tentative.html

<!DOCTYPE html>
<meta charset="utf-8" />
<title>Popup focus behaviors</title>
<link rel="author" href="mailto:masonf@chromium.org">
<link rel=help href="https://open-ui.org/components/popup.research.explainer">
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="resources/popup-utils.js"></script>

<div popup data-test='default behavior - popup is not focused' data-no-focus>
  <p>This is a popup</p>
  <button>first button</button>
</div>

<div popup data-test='autofocus popup' autofocus tabindex=-1 class=should-be-focused>
  <p>This is a popup</p>
</div>

<div popup data-test='autofocus empty popup' autofocus tabindex=-1 class=should-be-focused></div>

<div popup data-test='autofocus popup with button' autofocus tabindex=-1 class=should-be-focused>
  <p>This is a popup</p>
  <button>button</button>
</div>

<div popup data-test='autofocus child'>
  <p>This is a popup</p>
  <button autofocus class=should-be-focused>autofocus button</button>
</div>

<div popup data-test='autofocus on tabindex=0 element'>
  <p autofocus tabindex=0 class=should-be-focused>This is a popup with autofocus on a tabindex=0 element</p>
  <button>button</button>
</div>

<div popup data-test='autofocus multiple children'>
  <p>This is a popup</p>
  <button autofocus class=should-be-focused>autofocus button</button>
  <button autofocus>second autofocus button</button>
</div>

<div popup autofocus tabindex=-1 data-test='autofocus popup and multiple autofocus children' class=should-be-focused>
  <p>This is a popup</p>
  <button autofocus>autofocus button</button>
  <button autofocus>second autofocus button</button>
</div>

<div popup delegatesfocus data-test='delegatesfocus popup'>
  <p>This is a popup</p>
  <button class=should-be-focused>first button should be focused</button>
  <button>second button</button>
</div>

<div popup delegatesfocus data-test='delegatesfocus takes precedence over autofocus'>
  <p>This is a popup</p>
  <button class=should-be-focused>first button</button>
  <button autofocus>autofocus button should NOT be focused</button>
</div>

<div popup delegatesfocus autofocus data-test='delegatesfocus takes precedence over autofocus 2'>
  <p>This is a popup</p>
  <button class=should-be-focused>first button</button>
  <button>autofocus button should NOT be focused</button>
</div>

<div popup delegatesfocus data-test='delegatesfocus on empty popup has no effect' data-no-focus></div>

<div popup data-test='delegatesfocus on child has no effect' data-no-focus>
  <p>This is a popup</p>
  <button delegatesfocus>first button</button>
</div>

<div popup delegatesfocus data-test='delegatesfocus skips contained popups'>
  <p>This is a popup</p>
  <div popup>
    <button>Contained popup button</button>
  </div>
  <button class=should-be-focused>first button</button>
  <button>autofocus button should NOT be focused</button>
</div>

<div popup delegatesfocus data-test='delegatesfocus skips contained dialogs'>
  <p>This is a popup</p>
  <dialog>
    <button>Contained dialog button</button>
  </dialog>
  <button class=should-be-focused>first button</button>
  <button>autofocus button should NOT be focused</button>
</div>

<style>
  [popup] {
    border: 2px solid black;
    top:150px;
    left:150px;
  }
  :focus-within { border: 5px dashed red; }
  :focus { border: 5px solid lime; }
</style>

<script>
  function addInvoker(t, popup) {
    const button = document.createElement('button');
    button.innerText = 'Click me';
    const popupId = 'popup-id';
    assert_equals(document.querySelectorAll('#' + popupId).length, 0);
    document.body.appendChild(button);
    t.add_cleanup(function() {
      popup.removeAttribute('id');
      button.remove();
    });
    popup.id = popupId;
    button.setAttribute('togglepopup', popupId);
    return button;
  }
  function addPriorFocus(t) {
    const priorFocus = document.createElement('button');
    document.body.appendChild(priorFocus);
    t.add_cleanup(() => priorFocus.remove());
    return priorFocus;
  }
  function activateAndVerify(popup) {
    const testName = popup.getAttribute('data-test');
    promise_test(async t => {
      const priorFocus = addPriorFocus(t);
      let expectedFocusedElement = popup.matches('.should-be-focused') ? popup : popup.querySelector('.should-be-focused');
      if (popup.hasAttribute('data-no-focus')) {
        expectedFocusedElement = priorFocus;
      }
      assert_true(!!expectedFocusedElement);
      assert_false(popup.matches(':top-layer'));

      // Directly show and hide the popup:
      priorFocus.focus();
      assert_equals(document.activeElement, priorFocus);
      popup.showPopup();
      assert_equals(document.activeElement, expectedFocusedElement, `${testName} activated by popup.showPopup()`);
      popup.hidePopup();
      assert_equals(document.activeElement, priorFocus, 'prior element should get focus on hide');

      // Hit Escape:
      priorFocus.focus();
      assert_equals(document.activeElement, priorFocus);
      popup.showPopup();
      assert_equals(document.activeElement, expectedFocusedElement, `${testName} activated by popup.showPopup()`);
      await sendEscape();
      assert_equals(document.activeElement, priorFocus, 'prior element should get focus after Escape');

      // Change the popup type:
      priorFocus.focus();
      popup.showPopup();
      assert_equals(document.activeElement, expectedFocusedElement, `${testName} activated by popup.showPopup()`);
      assert_equals(popup.popup, 'auto', 'All popups in this test should start as popup=auto');
      popup.popup = 'hint';
      assert_false(popup.matches(':top-layer'), 'Changing the popup type should hide the popup');
      assert_equals(document.activeElement, priorFocus, 'prior element should get focus when the type is changed');
      popup.popup = 'auto';

      // Remove from the document:
      priorFocus.focus();
      popup.showPopup();
      assert_equals(document.activeElement, expectedFocusedElement, `${testName} activated by popup.showPopup()`);
      popup.remove();
      assert_false(popup.matches(':top-layer'), 'Removing the popup should hide it');
      if (!popup.hasAttribute('data-no-focus')) {
        assert_not_equals(document.activeElement, priorFocus, 'prior element should *not* get focus when the popup is removed from the document');
      }
      document.body.appendChild(popup);

      // Show a modal dialog:
      priorFocus.focus();
      popup.showPopup();
      assert_equals(document.activeElement, expectedFocusedElement, `${testName} activated by popup.showPopup()`);
      const dialog = document.body.appendChild(document.createElement('dialog'));
      dialog.showModal();
      assert_false(popup.matches(':top-layer'), 'Opening a modal dialog should hide the popup');
      assert_not_equals(document.activeElement, priorFocus, 'prior element should *not* get focus when a modal dialog is shown');
      dialog.close();
      dialog.remove();

      // Use an activating element:
      const button = addInvoker(t, popup);
      priorFocus.focus();
      button.click();
      assert_true(popup.matches(':top-layer'));
      assert_equals(document.activeElement, expectedFocusedElement, `${testName} activated by button.click()`);

      // Make sure we can directly focus the (already open) popup:
      popup.focus();
      assert_equals(document.activeElement, popup.hasAttribute('tabindex') ? popup : expectedFocusedElement, `${testName} directly focus with popup.focus()`);
      button.click(); // Button is set to toggle the popup
      assert_false(popup.matches(':top-layer'));
      assert_equals(document.activeElement, priorFocus, 'prior element should get focus on button-toggled hide');
    }, "Popup focus test: " + testName);

    promise_test(async t => {
      const priorFocus = addPriorFocus(t);
      assert_false(popup.matches(':top-layer'), 'popup should start out hidden');
      let button = addInvoker(t, popup);
      assert_equals(button.getAttribute('togglepopup'), popup.id, 'This test assumes the button uses `togglepopup`.');
      assert_not_equals(button, priorFocus, 'Stranger things have happened');
      assert_false(popup.contains(button), 'Start with a non-contained button');
      priorFocus.focus();
      assert_equals(document.activeElement, priorFocus);
      popup.showPopup();
      assert_true(popup.matches(':top-layer'));
      await clickOn(button); // This will not light dismiss, but will "toggle" the popup.
      assert_false(popup.matches(':top-layer'));
      assert_equals(document.activeElement, priorFocus, 'Focus should return to prior focus');

      // Same thing, but the button is contained within the popup
      button.removeAttribute('togglepopup');
      button.setAttribute('hidepopup', popup.id);
      popup.appendChild(button);
      priorFocus.focus();
      popup.showPopup();
      assert_true(popup.matches(':top-layer'));
      if (!popup.hasAttribute('data-no-focus')) {
        assert_not_equals(document.activeElement, priorFocus, 'focus should shift for this element');
      }
      await clickOn(button);
      assert_false(popup.matches(':top-layer'), 'clicking button should hide the popup');
      assert_equals(document.activeElement, priorFocus, 'Contained button should return focus to the previously focused element');

      // Same thing, but the button is unrelated (no togglepopup)
      button = document.createElement('button');
      document.body.appendChild(button);
      priorFocus.focus();
      popup.showPopup();
      assert_true(popup.matches(':top-layer'));
      await clickOn(button); // This will light dismiss the popup, focus the prior focus, then focus this button.
      assert_false(popup.matches(':top-layer'), 'clicking button should hide the popup (via light dismiss)');
      assert_equals(document.activeElement, button, 'Focus should go to unrelated button on light dismiss');
    }, "Popup button click focus test: " + testName);

    promise_test(async t => {
      if (popup.hasAttribute('data-no-focus')) {
        // This test only applies if the popup changes focus
        return;
      }
      const priorFocus = addPriorFocus(t);
      assert_false(popup.matches(':top-layer'), 'popup should start out hidden');

      // Move the prior focus out of the document
      priorFocus.focus();
      popup.showPopup();
      assert_true(popup.matches(':top-layer'));
      const newFocus = document.activeElement;
      assert_not_equals(newFocus, priorFocus, 'focus should shift for this element');
      priorFocus.remove();
      assert_equals(document.activeElement, newFocus, 'focus should not change when prior focus is removed');
      popup.hidePopup();
      assert_not_equals(document.activeElement, priorFocus, 'focused element has been removed');
      document.body.appendChild(priorFocus); // Put it back

      // Move the prior focus inside the (already open) popup
      priorFocus.focus();
      popup.showPopup();
      assert_true(popup.matches(':top-layer'));
      assert_false(popup.contains(priorFocus), 'Start with a non-contained prior focus');
      popup.appendChild(priorFocus); // Move inside the popup
      assert_true(popup.contains(priorFocus));
      assert_true(popup.matches(':top-layer'), 'popup should stay open');
      popup.hidePopup();
      assert_not_equals(document.activeElement, priorFocus, 'focused element is display:none inside the popup');
      document.body.appendChild(priorFocus); // Put it back
    }, "Popup corner cases test: " + testName);
  }

  document.querySelectorAll('body > [popup]').forEach(popup => activateAndVerify(popup));
</script>
